+++ programs/steward/src/instructions/update_parameters.rs
@@ -17,8 +17,14 @@ pub fn handler(
     let mut config = ctx.accounts.config.load_mut()?;
     let max_slots_in_epoch = EpochSchedule::get()?.slots_per_epoch;
     let current_epoch = Clock::get()?.epoch;
-    config
-        .parameters
-        .update(update_parameters_args, current_epoch, max_slots_in_epoch)?;
+
+    let new_parameters = config.parameters.get_updated_parameters(
+        update_parameters_args,
+        current_epoch,
+        max_slots_in_epoch,
+    )?;
+
+    config.parameters = new_parameters;
+
     Ok(())
 }
