+++ tests/tests/steward/mod.rs
@@ -1,5 +1,6 @@
 mod test_algorithms;
 mod test_integration;
+mod test_parameters;
 mod test_spl_passthrough;
 mod test_state_methods;
 mod test_state_transitions;
